intelligence-invoice-engine/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
pillow
sentence-transformers
chromadb
pydantic
python-multipart
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters
langchain-groq
langgraph
groq
python-dotenv
pandas
openpyxl
xlrd==1.2.0
requests